% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concord}
\alias{concord}
\title{concordance analysis}
\usage{
concord(x, y, ncomp = 2, dmod = 1, center = TRUE, scale = FALSE,
  option = "uniform", kx = "all", ky = "all", wx = 1, wy = 1,
  pos = FALSE, verbose = TRUE)
}
\arguments{
\item{x}{a list of predictive matrices. The columns are observations, rows are varaibles. The columns (observations)
has to be matched.}

\item{y}{a response matrix. Rows are variables, columns are observations. The columns should be matched with columns in x.}

\item{ncomp}{the number of components want to retain}

\item{dmod}{the deflation mode, dmod = 2 is the original publication of}

\item{center}{logical values, whether the variables should be centered}

\item{scale}{logical values, whether the variables should be scaled}

\item{option}{the option for normalizing matrix}

\item{kx}{the}

\item{ky}{the number (if it is an integer > 1) or the proportion (if 0 < ky < 1) of}

\item{wx}{weight for the rows of x}

\item{wy}{weight for the rows of y}

\item{pos}{logical value, whether only non-negative loadings retained}

\item{verbose}{if the process of calculation should be printed}
}
\description{
concordance analysis
}
\author{
Chen Meng
}

