% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.concord.R
\name{cv.concord}
\alias{cv.concord}
\title{Cross-validation for concordance analysis}
\usage{
cv.concord(x, y, fold = 7, opt.kx = seq(0.1, 0.6, length.out = 10),
  opt.ky = seq(0.1, 0.9, length.out = 10), ...)
}
\arguments{
\item{x}{A list of predictive matrices}

\item{y}{the dependent matrix}

\item{fold}{the number of fold for cross validation}

\item{opt.kx}{a numeric vector, candidate kx to be evaluated}

\item{opt.ky}{a numeric vector, candidate ky to be evaluated}

\item{...}{other arguments to be passed to concord}
}
\value{
A list of two data.frames. One for x and one for y. Each data.frame has three columns:
  kx/ky (input values of opt.kx/opt.ky), mean - the cv (cross-validation) mean, sd - the sd
  of cv-error
}
\description{
Cross-validation for concordance analysis to find optimal sparsity level of x and y
}
\examples{

}
\author{
Chen Meng
}
