% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.softSVD.R
\name{cv.softKv}
\alias{cv.softKv}
\title{cross-validation for softSVD}
\usage{
cv.softKv(x, nf = 1, kv.opt = c(0.3, 0.5, 0.8), wv = 1, wu = 1,
  pos = FALSE, maxiter = 50, tol = sqrt(.Machine$double.eps),
  verbose = FALSE, ncores = 1, fold = 5, nstart = 1, seed = NULL,
  loorss = FALSE)
}
\arguments{
\item{x}{input matrix}

\item{nf}{number of component}

\item{kv.opt}{optional value for sparsity on right singular value}

\item{wv}{weight for columns}

\item{wu}{weight for rows}

\item{pos}{whether retein non-negative results}

\item{maxiter}{maximum number of iteration}

\item{tol}{convergence tolerance}

\item{verbose}{if print the progress}

\item{ncores}{the number of cores used, passed to mclapply}

\item{fold}{fold number in cross validation}

\item{nstart}{how many time the k-fold cross validation should be done}

\item{seed}{set seed}

\item{loorss}{if the Leave-one-out procedure should be used in matrix reconstruction}
}
\description{
This function use k-fold cross-valiation method to optimize the sparsity
  of right singular values
}
\seealso{
\code{\link{cv.softSVD}}
}
