% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.softSVD.R
\name{cv.softSVD}
\alias{cv.softSVD}
\title{cross-validation for softSVD}
\usage{
cv.softSVD(x, nf = 1, kv.opt = c(0.3, 0.5, 0.8), ku.opt = c(0.3, 0.5,
  0.8), wv = 1, wu = 1, pos = FALSE, maxiter = 50,
  tol = sqrt(.Machine$double.eps), verbose = FALSE, ncores = 1,
  fold = 5, nstart = 1, seed = NULL, loorss = FALSE)
}
\arguments{
\item{x}{input matrix}

\item{nf}{number of component}

\item{kv.opt}{optional value for sparsity on right singular value}

\item{ku.opt}{optional value for sparsity on left singular value}

\item{wv}{weight for columns}

\item{wu}{weight for rows}

\item{pos}{whether retein non-negative results}

\item{tol}{convergence tolerance}

\item{verbose}{if print the progress}

\item{ncores}{the number of cores used, passed to mclapply}

\item{fold}{fold number in cross validation}

\item{nstart}{how many time the k-fold cross validation should be done}

\item{seed}{set seed}

\item{loorss}{if the Leave-one-out procedure should be used in matrix reconstruction}

\item{maiter}{maximum number of iteration}
}
\value{
list consist of two matrix. 
 - cvv - The PRESS for right singular vector
 - cvu - the PRESS for left singular vector
}
\description{
This function use k-fold cross-valiation method to optimize the sparsity
  of right singular values
}
\examples{
###
seed <- rnorm(8)
noise1 <- matrix(rnorm(8*15, sd = 0.3), 15, 8)
noise2 <- matrix(rnorm(8*20, sd = 0.3), 8, 20)
noise1[1:4, ] <- noise1[1:4, ] + rbind(seed, seed, seed, seed)
noise2[, 1:6] <- noise2[, 1:6] + seed
mat <- noise1 \%*\% noise2
dim(mat)

cv <- cv.softSVD(x = mat, nf = 1, kv.opt = 1:10, ku.opt = 1:8)
boxplot(cv$cvu)
boxplot(cv$cvv)
}
\keyword{internal}
