% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nci60_4arrays.R
\docType{data}
\name{NCI60_4arrays}
\alias{NCI60_4arrays}
\title{NCI60 microarray example data}
\format{a list of 4 \code{data.frame}s}
\usage{
NCI60_4arrays
}
\description{
A list of data.frame store the gene expression of NCI60 cell lines.
}
\keyword{datasets}

