% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normvec.R
\name{normvec}
\alias{normvec}
\title{Normalized a vector to unit length}
\usage{
normvec(x, center = FALSE)
}
\arguments{
\item{x}{a vector or matrix has one row/column}

\item{center}{whether x should be centered}
}
\value{
a normalized vector/matrix has the same length as x
}
\description{
Normalized a vector or matrix with 1 row/column to unit length
}
\author{
Chen Meng
}
\keyword{internal}
