% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.concord.R
\name{predictConcordance}
\alias{predictConcordance}
\title{predict y using result of concordance analysis}
\usage{
predictConcordance(object, newdata)
}
\arguments{
\item{object}{an object returned by concord}

\item{newdata}{new x}
}
\value{
the predicted y matrix
}
\description{
predict y using the result of concordance analysis
}
\author{
Chen Meng
}
