% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processOpt.R
\name{processOpt}
\alias{processOpt}
\title{Preprocess a list of matrices for concordance analysis}
\usage{
processOpt(x, center = TRUE, scale = FALSE, option = c("lambda1",
  "inertia", "uniform", "nrow"), value = 1)
}
\arguments{
\item{x}{a list of matrices to be preprocessed. The columns are variables and rows are observations}

\item{center}{if the variables should be centered}

\item{scale}{if the variables should be scaled}

\item{option}{options for normalizing matrices, possible value are "lambda1", "inertia", "uniform", "nrow"}

\item{value}{Only used if option is "lambda1" or inertia, the matrix are weighted by value/lambda1 
(or inertia)}
}
\value{
a list of preprocess matrices has the same dimension as input x
}
\description{
Internal function used to normalize a list of matrices passed into concordance analysis
}
\examples{
data("NCI60_4arrays")
dat <- lapply(NCI60_4arrays, t)
v <- processOpt(dat, option = "lambda1")

}
\author{
Chen Meng
}
\keyword{internal}
