% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sconcord.R
\name{sconcord}
\alias{sconcord}
\title{Sparse concordance analysis with cross-validation}
\usage{
sconcord(x, y, opt.kx, opt.ky, fold = 5, ncores = 1, ncomp = 1,
  dmod = 1, center = TRUE, scale = FALSE, option = "uniform", wx = 1,
  wy = 1, pos = FALSE)
}
\arguments{
\item{x}{A list of predictive matrices}

\item{y}{the dependent matrix}

\item{opt.kx}{a numeric vector, candidate kx to be evaluated}

\item{opt.ky}{a numeric vector, candidate ky to be evaluated}

\item{fold}{the number of fold for cross validation}

\item{ncores}{the number of cores to be used, passed to \code{mclapply} in "parallel" package.}

\item{ncomp}{see \code{\link{concord}}}

\item{dmod}{see \code{\link{concord}}}

\item{center}{see \code{\link{concord}}}

\item{scale}{see \code{\link{concord}}}

\item{option}{see \code{\link{concord}}}

\item{wx}{see \code{\link{concord}}}

\item{wy}{see \code{\link{concord}}}

\item{pos}{see \code{\link{concord}}}

\item{kx}{see \code{\link{concord}}}

\item{ky}{see \code{\link{concord}}}
}
\value{
concordance with results of optimization of kx, ky params
}
\description{
Sparse concordance analysis, for each component, the optimal sparsity level is 
  optimized using cross-validation
}
\examples{
# example here

}
\author{
Chen Meng
}
