% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softK.R
\name{softK}
\alias{softK}
\title{weighted soft thresholding a vector}
\usage{
softK(x, k, w = 1, pos = FALSE)
}
\arguments{
\item{x}{a vector}

\item{k}{the number (if integer k>1) or proportion (if 0<k<1) of non-zero elements}

\item{w}{the weight for each element, see "detail" section}

\item{pos}{if only retain non-negative values}
}
\description{
soft-thresholding a vector with possibility of using weight and non-negative constraint
}
\details{
using weight w
The vector x*w is used to determine which value to keep, but the values in x are kept. For example,
the vector x is c(1, 2, 3, 4, 5), w is c(2.5, 2.5, 2.5, 1, 1), the weighted vector is 
c(2.5, 5, 7.5, 4, 5). If k = 3, value in x corresponding to 5, 7.5 and 5 (i.e. 2, 3, 5) will be selected,
after thresholding, the values becomes 1, 2, 4.
}
\examples{
 
softK(1:5, w=c(2.5, 2.5, 2.5, 1, 1), k = 3)
}
\author{
Chen Meng
}
\keyword{internal}
