% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/targetResolve.R
\name{targetResolve}
\alias{targetResolve}
\title{resolving target space of drugs}
\usage{
targetResolve(t, s, m, ncomp = 1, center = TRUE, scale = TRUE,
  option = "uniform", kt = "all", km = "all", wt = 1, wm = 1,
  pos = FALSE, ncores = 1, fold = 5, init = "average", nstart = 1,
  seed = NULL, loorss = FALSE, scan = FALSE, nsd = 1, verbose = TRUE)
}
\arguments{
\item{t}{target matrix in the form target (rows) x drug (columns), values should be the higher the more potent, e.g. pKD.}

\item{s}{the sensitivity matrix, in the form drug (rows) x cell line (columns), the higher the more sensitive (e.g. -log10(GI50))}

\item{m}{a list of matrices, abundance of molecular features in the form of variables (rows) x cell lines (columns)}

\item{ncomp}{the number of components want to retain}

\item{center}{logical values, whether the variables should be centered}

\item{scale}{logical values, whether the variables should be scaled}

\item{option}{the option for normalizing matrix}

\item{kt}{the number (if it is an integer >= 1) or the proportion (if 0 < kt < 1) of kept variables in t. It should be 
a numeric value.}

\item{km}{the number (if it is an integer >= 1) or the proportion (if 0 < km < 1) of kept variables in m. It should be
a numeric value.}

\item{wt}{weight for the rows of t}

\item{wm}{weight for the rows of m}

\item{pos}{logical value, whether only non-negative loadings retained}

\item{ncores}{number of cores to be used, passed to \code{\link{mclapply}}}

\item{fold}{the number of fold to be used in cross-validation, only used if kx or ky is a vector}

\item{init}{how to initialize the algorithm. if no sparsity, svd is fast.}

\item{nstart}{how many time the k-fold cross validation should be done}

\item{seed}{set seed for random number generation}

\item{loorss}{if the Leave-one-out procedure should be used in matrix reconstruction}

\item{scan}{If the PRESS plot should be shown and used to determine the optimal k in CV}

\item{nsd}{the the n*sd for selecting k automatically}

\item{verbose}{if the process of calculation should be printed}
}
\description{
resolving target space and identify biomarkers by integrating drug target, cell line sensitivity and 
  cell line molecular data
}
\author{
Chen Meng
}
