% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concordance.R
\name{concord}
\alias{concord}
\title{concordance analysis}
\usage{
concord(x, y, ncomp = 2, dmod = 1, center.x = TRUE, scale.x = FALSE,
  center.y = TRUE, scale.y = FALSE, option = "uniform", kx = "all",
  ky = "all", wx = 1, wy = 1, pos = FALSE, verbose = TRUE,
  init = c("svd", "average")[2], ncores = 1, fold = 5, nstart = 1,
  seed = NULL, loorss = FALSE, scan = TRUE, nsd = 1)
}
\arguments{
\item{x}{a list of predictive matrices. The columns are observations, rows are varaibles. The columns (observations)
has to be matched.}

\item{y}{a response matrix. Rows are variables, columns are observations. The columns should be matched with columns in x.}

\item{ncomp}{the number of components want to retain}

\item{dmod}{the deflation mode, dmod = 2 is the original publication of}

\item{center.x}{logical values, whether the variables in x should be centered}

\item{scale.x}{logical values, whether the variables in x should be scaled}

\item{center.y}{logical values, whether the variables in y should be centered}

\item{scale.y}{logical values, whether the variables in y should be scaled}

\item{option}{the option for normalizing matrix}

\item{kx}{the number (if it is an integer > 1) or the proportion (if 0 < ky < 1) of kept variables in x. It should be
a numeric value.}

\item{ky}{the number (if it is an integer > 1) or the proportion (if 0 < ky < 1) of kept variables in y. It should be 
a numeric value.}

\item{wx}{weight for the rows of x}

\item{wy}{weight for the rows of y}

\item{pos}{logical value, whether only non-negative loadings retained}

\item{verbose}{if the process of calculation should be printed}

\item{init}{how to initialize the algorithm. if no sparsity, svd is fast.}

\item{ncores}{number of cores to be used, passed to \code{\link{mclapply}}}

\item{fold}{the number of fold to be used in cross-validation, only used if kx or ky is a vector}

\item{nstart}{how many time the k-fold cross validation should be done}

\item{seed}{set seed for random number generation}

\item{loorss}{if the Leave-one-out procedure should be used in matrix reconstruction}

\item{scan}{If the PRESS plot should be shown and used to determine the optimal k in CV}

\item{nsd}{the the n*sd for selecting k automatically}
}
\description{
concordance analysis
}
\examples{
library(omic3plus)
data("NCI60_4arrays")

y <- as.matrix(NCI60_4arrays$agilent)
x <- lapply(NCI60_4arrays[2:4], as.matrix)

# no sparsity
con1 <- concord(x, y, ncomp = 3)

# sparsity on rows of x, select 10\% genes
con <- concord(x, y, ncomp = 3, kx = 0.1)

# sparsity on rows of both x and y, select 10\% genes
con <- concord(x, y, ncomp = 3, kx = 0.1, ky = 0.1, option = "nk")
plot(con$score.x[, 1], con$score.y[, 1])
abline(a = 0, b = 1)
}
\author{
Chen Meng
}
